+++ Assets/_Scripts/Net/LocalPlayer.cs
@@ -108,7 +108,7 @@ public sealed class LocalPlayer : MonoBehaviour
         }
         if (selfGhost != null)   // self effect tint + hurt anim from the predicted mask (0 when not predicting, so they reset on leave)
         {
-            byte mask = prediction.Active ? StatusLogic.ActiveMask(prediction.Status) : (byte)0;
+            ushort mask = prediction.Active ? StatusLogic.ActiveMask(prediction.Status) : (ushort)0;
             var sv = selfGhost.GetComponent<StatusView>(); if (sv != null) sv.Render(mask);
             var fx = selfGhost.GetComponent<StatusFxView>(); if (fx != null) fx.Render(mask);
             var dv = selfGhost.GetComponent<DmgView>(); if (dv != null) dv.Render(mask);   // after StatusView (tint) — DmgView wins the hurt sprite
